generate_quote_images/requirements.txt
pillow
pytest-watch